requirements.txt
numpy
pyyaml
torch>=2.0
torchvision
scipy
tqdm